+++ angular_router/lib/src/rules/route_paths/route_path.dart
@@ -1,4 +1,4 @@
-import "../../url_parser.dart" show Url;
+import '../../url_parser.dart' show Url;
 
 class MatchedUrl {
   final String urlPath;