+++ angular_router/lib/src/utils.dart
@@ -1,8 +1,6 @@
-import "package:angular/src/core/linker/component_factory.dart"
-    show ComponentFactory;
-import "package:angular/src/core/reflection/reflection.dart" show reflector;
+import 'package:angular/angular.dart' show ComponentFactory, reflector;
 
-import "lifecycle/lifecycle_annotations.dart" show CanActivate;
+import 'lifecycle/lifecycle_annotations.dart' show CanActivate;
 
 class TouchMap {
   Map<String, String> map = {};