+++ angular/CHANGELOG.md
@@ -34,6 +34,8 @@ class AdBannerComponent implements AfterViewInit {
   `angular/angular.dart`.
 * Removed `angular/platform/common_dom.dart`; replace imports with
   `angular/angular.dart`.
+* Router code extracted into new 'angular_router' package.  Replace
+  'angular/router.dart' imports with 'angular_router/angular_router.dart'.
 
 ## 4.0.0-alpha+1